core router: make strategy a folder (one module per strategy, simple_shuffle)
